+++ tools/code_processing/loader.py
@@ -1,6 +1,5 @@
-import ast
-
-from . import python_to_ts, pseudocode, cpp
+from .python_source import SourceCode
+from .langs import ts, pseudocode, cpp
 
 
 language_names = {
@@ -11,17 +10,12 @@ language_names = {
 }
 
 
-def code_to_ast(source):
-    return ast.parse(source, type_comments=True)
-
-
 def code_to_samples(source):
-    tree = code_to_ast(source)
-    comments = python_to_ts.CommentData(source)
+    data = SourceCode(source)
     return {
-        "ast": tree,
-        "pseudo": pseudocode.PseudoCode().convert(tree, comments),
+        "ast": data.ast,
+        "pseudo": pseudocode.PseudoCode().convert(data),
         "py": source,
-        "cpp": cpp.CppTranslator().convert(tree, comments),
-        "ts": python_to_ts.Py2Ts().convert(tree, comments),
+        "cpp": cpp.CppTranslator().convert(data),
+        "ts": ts.TypeScriptTranslator().convert(data),
     }
